+++ core/exceptions/__init__.py
@@ -1,3 +1,0 @@
-from core.exceptions.base import AICError, ConfigurationError
-
-__all__ = ["AICError", "ConfigurationError"]